docs(samples): Moved get_project_number into test_utilities.py
